+++ python/src/lib.rs
@@ -31,6 +31,7 @@ mod tokenizer;
 #[pymodule]
 fn sudachipy(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     m.add_class::<dictionary::PyDictionary>()?;
+    m.add_class::<dictionary::PyDictionaryEntryIterator>()?;
     m.add_class::<tokenizer::PySplitMode>()?;
     m.add_class::<tokenizer::PyTokenizer>()?;
     m.add_class::<morpheme::PyMorphemeListWrapper>()?;