+++ sudachi/src/dic/word_info.rs
@@ -23,7 +23,8 @@ mod raw;
 
 pub(crate) use binary::{parse_i32_array, parse_u32_array, parse_user_data};
 pub use data::*;
-pub use infos::WordInfos;
+pub(crate) use infos::WordInfoEntryIdCursor;
+pub use infos::{WordInfoError, WordInfos};
 pub(crate) use layout::WordInfoVariableLayout;
 pub use parse::WordInfoParser;
 pub(crate) use raw::WordInfoVariableData;